chore: move create-validation-middleware
