+++ src/validator/index.ts
@@ -2,4 +2,4 @@ export * from '@/validator/factory/create-validator';
 export * from './types';
 export * from './constraints';
 export * from './flatten-violations';
-export { createValidationMiddleware } from './create-validation-middleware';
+export { createValidationMiddleware } from './middleware/create-validation-middleware';
